Refactored schema generation
Also introduced PhpSpec, plus a lot of specs for the items below.

The schema generator iterates on domain data using the registered iterators. They send each set of variables
to the workers. It also initializes the schema, and passes it on to iterators and schema initializer workers for further initialization.
It is used by the command to generate the schema.

Data used used for generation is provided by Schema\Domain\Iterator objects. They replace the former "Builders",
that both iterated and worked on data.

Workers are now generalized, and are used by all domains. Instead of manipulating an array passed by reference,
they receive a SchemaBuilder object.

The SchemaBuilder Provides simple functions to add and check for the existence of schema items.
Uses input objects, from the Schema\Builder\Input domain, to define items (Type, Field, Arg...).

Schema generation has also been restructured. Schema is the root namespace, with the common interfaces.
The Schema\Domain subnamespace has the Iterator service, as well as the included domains: Content and ImagesVariations.